initi clamped table wrapper
